fix: new signatures are effectively available only after reload
