+++ src/views/settings/signature-settings.tsx
@@ -15,19 +15,16 @@ import {
 	Button,
 	Padding,
 	SelectItem,
-	ButtonProps,
-	useSnackbar
+	ButtonProps
 } from '@zextras/carbonio-design-system';
 import { t, useIntegratedComponent } from '@zextras/carbonio-shell-ui';
-import { map, unescape, reject, concat } from 'lodash';
+import { map, reject, concat } from 'lodash';
 import styled from 'styled-components';
 import { v4 as uuidv4 } from 'uuid';
 
 import SelectIdentitySignature from './components/select-identity-signature';
 import { signaturesSubSection, setDefaultSignaturesSubSection } from './subsections';
-import { TIMEOUTS } from '../../constants';
 import { NO_SIGNATURE_ID, NO_SIGNATURE_LABEL } from '../../helpers/signatures';
-import { GetAllSignatures } from '../../store/actions/signatures';
 import type { SignatureSettingsPropsType, SignItemType } from '../../types';
 
 const DeleteButton = styled(Button)`
@@ -67,8 +64,7 @@ const SignatureSettings: FC<SignatureSettingsPropsType> = ({
 	updateIdentities,
 	setDisabled,
 	signatures,
-	setSignatures,
-	setOriginalSignatures
+	setSignatures
 }): ReactElement => {
 	const [currentSignature, setCurrentSignature] = useState<SignItemType | undefined>(undefined);
 	const [Composer, composerIsAvailable] = useIntegratedComponent('composer');
@@ -83,52 +79,6 @@ const SignatureSettings: FC<SignatureSettingsPropsType> = ({
 	const setEditor = (editor: EditorType): void => {
 		editorRef.current.editor = editor;
 	};
-	const createSnackbar = useSnackbar();
-
-	const onSignaturesLoaded = useCallback(
-		(signs: Array<SignItemType>) => {
-			const signaturesItems = map(
-				signs,
-				(item: SignItemType, idx) =>
-					({
-						label: item.name,
-						name: item.name,
-						id: item.id,
-						description: unescape(item?.content?.[0]?._content)
-					}) as SignItemType
-			);
-			setSignatures(signaturesItems);
-			setOriginalSignatures(
-				signaturesItems.map((el) => ({
-					id: el.id,
-					name: el.label ?? '',
-					label: el.label ?? '',
-					description: el.description ?? ''
-				}))
-			);
-		},
-		[setOriginalSignatures, setSignatures]
-	);
-
-	// Fetches signatures from the BE
-	useEffect(() => {
-		GetAllSignatures()
-			.then(({ signature: signs }) => onSignaturesLoaded(signs))
-			.catch((err) => {
-				console.error(err);
-				createSnackbar({
-					key: `get-all-signature`,
-					replace: true,
-					type: 'error',
-					label: t(
-						'label.error_unable_load_signatures',
-						'Unable to load signatures. Please reload or try again later.'
-					),
-					autoHideTimeout: TIMEOUTS.SNACKBAR_DEFAULT_TIMEOUT,
-					hideButton: true
-				});
-			});
-	}, [createSnackbar, onSignaturesLoaded]);
 
 	// Set the default current signature if missing
 	useEffect(() => {
